Fix Create User Controller, Login Controller
